+++ CHANGES.md
@@ -7,6 +7,7 @@ New:
 - Added `self_sync` option to `input.srt` to accommodate for streams in file mode (#3684)
 - Added `curve` parameter to fade functions and `liq_fade_{in,skip,out}_curve` metadata override (#3691)
 - Added `delay` parameter to fade functions to make it possible to add delay before fade happens. Add `liq_fade_{in,skip,out}_delay` metadata override.
+- Added `single_track` option to allow `sequence` to play each source until they are unavailable while keeping track marks.
 
 Changed: